requirements.txt
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'

# AI providers (install one or both based on preference)
openai>=1.0.0
anthropic>=0.18.0
